Wed Jun 28 16:21:10 2000                        Michael Jennings (mej)
pkgtool is now up to the point where it can build RPM's either from
SRPM's or from a spec file.  It can also install and/or query for
contents RPM and DEB packages.  (The DEB build process is more
complex, so I'll wait on it.)

The source generation stuff is next.