Add Vagrantfile and .gitignore
The new Vagrantfile will build and launch a CentOS 7-based VM with
Mezzanine installed simply by running `vagrant up` in the top-level
directory.  You can then connect to this VM using `vagrant ssh` and
play around with the various Mezzanine commands.  See
https://docs.vagrantup.com/ for more info on Vagrant.

A .gitignore file was also added to ignore the .vagrant/ directory
that Vagrant uses along with a few other file types we don't want to
commit to the repository.